third_party/WebKit/LayoutTests/media/track/regions-webvtt/vtt-region-list.html

<!DOCTYPE html>
<title>Tests VTTRegionList functionality: length, operator[], and getRegionById().</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var testTrack = document.createElement('track');
    
    assert_equals(testTrack.track.mode, 'disabled');
    assert_equals(testTrack.track.regions, null);

    testTrack.track.mode = 'hidden';
    var regions = testTrack.track.regions;

    assert_true(regions instanceof VTTRegionList, 'instanceof');

    assert_equals(regions.length, 0);

    var region = new VTTRegion();
    region.id = 'TestId';

    assert_equals(region.track, null);

    testTrack.track.addRegion(region);

    assert_equals(regions.length, 1);
    assert_equals(regions[0], region);
    assert_equals(regions[0].track, testTrack.track);

    assert_equals(region.track, testTrack.track);

    var updatedRegion = new VTTRegion();
    updatedRegion.id = region.id;
    updatedRegion.viewportAnchorX = 59;
    updatedRegion.viewportAnchorY = 68;
    updatedRegion.regionAnchorX = 20;
    updatedRegion.regionAnchorY = 30;
    updatedRegion.height = 5;
    updatedRegion.width = 87;
    updatedRegion.scroll = 'up';

    testTrack.track.addRegion(updatedRegion);
    assert_equals(regions[0].viewportAnchorX, updatedRegion.viewportAnchorX);
    assert_equals(regions[0].viewportAnchorY, updatedRegion.viewportAnchorY);
    assert_equals(regions[0].regionAnchorX, updatedRegion.regionAnchorX);
    assert_equals(regions[0].regionAnchorY, updatedRegion.regionAnchorY);
    assert_equals(regions[0].height, updatedRegion.height);
    assert_equals(regions[0].width, updatedRegion.width);
    assert_equals(regions[0].scroll, updatedRegion.scroll);

    assert_not_equals(regions[0], updatedRegion);

    testTrack.track.addRegion(region);
    assert_equals(regions.length, 1);
    testTrack.track.removeRegion(region);
    assert_equals(regions.length, 0);

    assert_throws('NotFoundError', function() { testTrack.track.removeRegion(region); });

    // FIXME(109818): Update test for multiple initial regions (after parsing is supported).
});
</script>
</html>
